+++ lib/core/di/di.dart
@@ -1,5 +1,6 @@
 import 'package:connectivity_plus/connectivity_plus.dart';
 import 'package:dio/dio.dart';
+import 'package:flutter/foundation.dart';
 import 'package:get_it/get_it.dart';
 import 'package:logger/logger.dart';
 
@@ -48,6 +49,31 @@ Future<void> setupDI() async {
 
   di.registerLazySingleton(() => dio);
   di.registerLazySingleton(() => ApiClient(di<Dio>()));
+
+  if (kDebugMode) {
+    final logger = di<AppLogger>();
+    dio.interceptors.add(
+      InterceptorsWrapper(
+        onRequest: (options, handler) {
+          logger.d('-> ${options.method} ${options.uri}');
+          return handler.next(options);
+        },
+        onResponse: (response, handler) {
+          logger.d(
+            '<- ${response.statusCode} ${response.requestOptions.method} ${response.requestOptions.uri}',
+          );
+          return handler.next(response);
+        },
+        onError: (DioException error, handler) {
+          final request = error.requestOptions;
+          logger.w(
+            'x ${error.response?.statusCode ?? '-'} ${error.type} ${request.method} ${request.uri} ${error.message ?? ''}',
+          );
+          return handler.next(error);
+        },
+      ),
+    );
+  }
 
   await di.allReady();
 }